+++ compose-highlight/src/androidTest/kotlin/dev/hossain/highlight/ui/RememberSyntaxHighlightedEditorValueTest.kt
@@ -3,7 +3,7 @@ package dev.hossain.highlight.ui
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.test.junit4.createComposeRule
+import androidx.compose.ui.test.junit4.v2.createComposeRule
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.test.ext.junit.runners.AndroidJUnit4